+++ frontend/src/components/TripActivitiesCard.jsx
