+++ frontend/src/pages/Trips.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import styles from '../styles/TripPlanner.module.css';
+import styles from '../styles/Trips.module.css';
 import TripCard from '../components/TripCard';
 
-const TripPlanner = () => {
+const Trips = () => {
   const trips = [
     {
       title: 'Bali Retreat',
@@ -36,7 +36,7 @@ const TripPlanner = () => {
   ];
 
   return (
-    <main className={styles.tripPlanner}>
+    <main className={styles.trips}>
       <header className={styles.header}>
         <img src="https://cdn.builder.io/api/v1/image/assets/TEMP/f7f79458-a1c8-4104-8c49-5cf47daaa824?placeholderIfAbsent=true&apiKey=3a931a869e0e4b4bb92008bc01989510" alt="Trip Planner Logo" className={styles.logo} />
         <nav className={styles.navigation}>
@@ -61,4 +61,4 @@ const TripPlanner = () => {
   );
 };
 
-export default TripPlanner;
+export default Trips;
