+++ frontend/src/styles/ActivityOverviewCard.module.css
